requirements.txt
matplotlib==3.1.0
numpy==1.16.4
Pillow==6.0.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
torch==1.1.0
torchvision==0.3.0
